Add explicit typecasting for some constants

+++ Telegram/SourceFiles/boxes/stickers_box.cpp
@@ -485,7 +485,11 @@ bool StickersBox::installFail(uint64_t setId, const RPCError &error) {
 void StickersBox::preloadArchivedSets() {
 	if (!_tabs) return;
 	if (!_archivedRequestId) {
-		_archivedRequestId = MTP::send(MTPmessages_GetArchivedStickers(MTP_flags(0), MTP_long(0), MTP_int(kArchivedLimitFirstRequest)), rpcDone(&StickersBox::getArchivedDone, 0ULL));
+		_archivedRequestId =
+				MTP::send(MTPmessages_GetArchivedStickers(MTP_flags(0),
+														  MTP_long(0),
+														  MTP_int(kArchivedLimitFirstRequest)),
+						  rpcDone(&StickersBox::getArchivedDone, uint64_t(0)));
 	}
 }
